First Commit
This is the initial commit.

So far we have an empty main scene, an empty main menu scene, and a functioning game over scene.

a set for the game over scene where the camera rotates around to view the set, and a menu for continuing or returning to the main menu