VictoryScreen+
added a victory screen, and updated various things on previous levels.

enabled the guy you picked up in earlier levels to carry over to the next, edited some of the attributes of the various guns.

added a deathplane for the third level, so if the player falls off the map they take damage and then are warpped back to the arena.